Invert tab colours to follow common design conventions
Swapping the active and inactive tab colours might be a good idea. Just to follow the common design convention of having the tab colour match the background of the currently selected tab.

Was finding there was some confusion when only 2 tabs were open. With more tabs open this change doesn't matter as much. Could be worth thinking about, even if this change doesn't make it upstream.